refactor: move helper module outside of app submodule
