+++ src/biomodals/app/bioinfo/gromacs_app.py
@@ -18,8 +18,8 @@ import modal
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MAX_TIMEOUT
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import run_command
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import run_command
 
 ##########################################
 # Modal configs
