+++ src/biomodals/app/design/boltzgen_app.py
@@ -18,8 +18,8 @@ import orjson
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MAX_TIMEOUT, MODEL_VOLUME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import (
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import (
     package_outputs,
     run_command,
     run_command_with_log,
