+++ src/biomodals/app/design/ligandmpnn_app.py
@@ -20,13 +20,13 @@ import modal
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MAX_TIMEOUT, MODEL_VOLUME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import (
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import (
     find_with_fd,
     package_outputs,
     run_command_with_log,
 )
-from biomodals.app.helper.web import download_files
+from biomodals.helper.web import download_files
 
 ##########################################
 # Modal configs
