+++ src/biomodals/app/design/ppiflow_app.py
@@ -12,8 +12,8 @@ from pydantic import BaseModel, computed_field, model_validator
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MAX_TIMEOUT, MODEL_VOLUME, MODEL_VOLUME_NAME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import run_command_with_log, sanitize_filename
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import run_command_with_log, sanitize_filename
 
 ##########################################
 # Modal configs
