+++ src/biomodals/app/fold/abcfold2_app.py
@@ -24,9 +24,9 @@ import modal
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MODEL_VOLUME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import package_outputs
-from biomodals.app.helper.web import download_files
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import package_outputs
+from biomodals.helper.web import download_files
 
 ##########################################
 # Modal configs
