+++ src/biomodals/app/score/abnativ_app.py
@@ -16,8 +16,8 @@ import modal
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MAX_TIMEOUT, MODEL_VOLUME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import package_outputs, run_command, softlink_dir
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import package_outputs, run_command, softlink_dir
 
 ##########################################
 # Modal configs
