+++ src/biomodals/app/score/af3score_app.py
@@ -24,8 +24,8 @@ import modal
 
 from biomodals.app.config import AppConfig
 from biomodals.app.constant import MODEL_VOLUME
-from biomodals.app.helper import patch_image_for_helper
-from biomodals.app.helper.shell import (
+from biomodals.helper import patch_image_for_helper
+from biomodals.helper.shell import (
     copy_files,
     run_command,
     run_command_with_log,
